+++ src/components/Widgets/FeedbackWidget/realm.js
@@ -1,36 +1,17 @@
 import React from 'react';
 import * as Realm from 'realm-web';
 import { isBrowser } from '../../../utils/is-browser';
+import { removeAllRealmUsersFromLocalStorage } from '../../../utils/realm-user-management';
 
 const APP_ID = 'feedbackwidgetv3-dgcsv';
 export const app = isBrowser ? Realm.App.getApp(APP_ID) : { auth: {} };
-
-/**
- * @param {object} storage
- * @returns The prefix for the storage key used by Realm for localStorage access
- */
-function parseStorageKey(storage) {
-  if (!storage.keyPart) {
-    return '';
-  }
-  const prefix = parseStorageKey(storage.storage);
-  return prefix + storage.keyPart + ':';
-}
 
 /**
  * Deletes localStorage data for all users
  */
 function deleteLocalStorageData() {
   const { allUsers } = app;
-  // The accessToken and refreshToken are automatically removed if invalid, but not the following keys
-  const keysToDelete = ['profile', 'providerType'];
-
-  Object.values(allUsers).forEach((user) => {
-    const storageKeyPrefix = parseStorageKey(user.storage);
-    keysToDelete.forEach((key) => {
-      localStorage.removeItem(storageKeyPrefix + key);
-    });
-  });
+  removeAllRealmUsersFromLocalStorage(allUsers);
 }
 
 // User Authentication & Management
